Module tidying
