+++ sklearn_numba_dpex/device/__init__.py
@@ -1,0 +1,3 @@
+from .device import DeviceParams
+
+__all__ = ["DeviceParams"]
