+++ sklearn_numba_dpex/device/tests/test_device_params.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 import dpctl
 import pytest
 
-from sklearn_numba_dpex.utils._device import _DeviceParams
+from sklearn_numba_dpex.device import DeviceParams
 
 
 def test_opencl_requirement():
@@ -28,7 +28,7 @@ def test_warnings_non_cl_device_params():
         max_work_group_size: int = 16
         name: str = "Fake Sycl Device Without OpenCL Support"
 
-    device_params = _DeviceParams(_FakeSyclDevice())
+    device_params = DeviceParams(_FakeSyclDevice())
     # Check that a warning is raised if a device is not detected by opencl
     with pytest.warns(RuntimeWarning):
         device_params.preferred_work_group_size_multiple
@@ -40,7 +40,7 @@ def test_warnings_non_cl_device_params():
 def test_no_warnings_cl_device_params():
     try:
         sycl_device = dpctl.SyclDevice("opencl")
-        device_params = _DeviceParams(sycl_device)
+        device_params = DeviceParams(sycl_device)
     except dpctl.SyclDeviceCreationError:
         pytest.xfail("No opencl SyclDevice available")
